requirements.txt
amqp==2.6.1
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==4.4.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.3
djangorestframework==3.13.1
drf-yasg==1.20.0
flower==0.9.7
humanize==4.0.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
kombu==4.6.11
MarkupSafe==2.1.1
packaging==21.3
prometheus-client==0.8.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
pyparsing==3.0.7
python-dotenv==0.20.0
pytz==2022.1
redis==3.5.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tornado==6.1
uritemplate==4.1.1
urllib3==1.26.9
vine==1.3.0
wcwidth==0.2.5
wrapt==1.14.0